fix(spline): fix rendering of spline
