+++ package.json
@@ -26,7 +26,7 @@
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-tabs": "^1.0.4",
     "@radix-ui/react-toast": "^1.1.5",
-    "@splinetool/react-spline": "^3.1.1",
+    "@splinetool/react-spline": "^4.0.0",
     "@splinetool/runtime": "^1.8.1",
     "@standard-crypto/farcaster-js": "^7.4.0",
     "@t3-oss/env-nextjs": "^0.2.2",
